Improve split routers: Health, GraphQL and Web
