+++ .env.dev
@@ -18,7 +18,7 @@ FORCE_SSL=false
 PORT=4000
 SECRET_KEY_BASE= # Generate secret with `mix phx.gen.secret`
 SESSION_KEY=elixir_boilerplate
-SIGNING_SALT= # Generate salt with `mix phx.gen.secret`
+SESSION_SIGNING_SALT= # Generate salt with `mix phx.gen.secret`
 
 # Database configuration
 # - Use `postgres://localhost/elixir_boilerplate_dev` if you have a local PostgreSQL server
