+++ config/test.exs
@@ -21,8 +21,6 @@ config :elixir_boilerplate, ElixirBoilerplateWeb.Endpoint,
   http: [port: 4001],
   server: false,
   secret_key_base: "G0ieeRljoXGzSDPRrYc2q4ADyNHCwxNOkw7YpPNMa+JgP9iGgJKT4K96Bw/Mf/pd",
-  session_key: "test",
-  signing_salt: "qh+vmMHsOqcjKF3TSSIsghwt2go48m2+IQ+kMTOB3BrSysSr7D4a21uAtt4yp4wn",
   static_url: [
     scheme: "https",
     host: "example.com",
@@ -33,6 +31,10 @@ config :elixir_boilerplate, ElixirBoilerplateWeb.Endpoint,
     host: "example.",
     port: "443"
   ]
+
+config :elixir_boilerplate, ElixirBoilerplateWeb.Router,
+  session_key: "test",
+  session_signing_salt: "qh+vmMHsOqcjKF3TSSIsghwt2go48m2+IQ+kMTOB3BrSysSr7D4a21uAtt4yp4wn"
 
 config :logger, level: :warn
