+++ lib/elixir_boilerplate_graphql/router.ex
@@ -1,27 +1,29 @@
 defmodule ElixirBoilerplateGraphQL.Router do
   use Plug.Router
-  use NewRelic.Transaction
 
-  @absinthe_configuration [
-    document_providers: {ElixirBoilerplateGraphQL, :document_providers},
-    json_codec: Jason,
-    schema: ElixirBoilerplateGraphQL.Schema
-  ]
+  defmodule GraphQL do
+    use Plug.Router
+    use NewRelic.Transaction
+
+    plug(:match)
+    plug(:dispatch)
+
+    forward("/",
+      to: Absinthe.Plug,
+      init_opts: [
+        document_providers: {ElixirBoilerplateGraphQL, :document_providers},
+        json_codec: Jason,
+        schema: ElixirBoilerplateGraphQL.Schema
+      ]
+    )
+  end
 
   plug(ElixirBoilerplateGraphQL.Plugs.Context)
 
   plug(:match)
   plug(:dispatch)
 
-  forward("/graphql",
-    to: Absinthe.Plug,
-    init_opts: @absinthe_configuration
-  )
-
-  forward("/graphiql",
-    to: Absinthe.Plug.GraphiQL,
-    init_opts: [interface: :playground] ++ @absinthe_configuration
-  )
+  forward("/graphql", to: GraphQL)
 
   match(_, do: conn)
 end
