+++ lib/elixir_boilerplate_health/router.ex
@@ -1,0 +1,31 @@
+defmodule ElixirBoilerplateHealth.Router do
+  use Plug.Router
+
+  defmodule Health do
+    use Plug.Router
+    use NewRelic.Transaction
+
+    plug(:match)
+    plug(:dispatch)
+
+    forward(
+      "/",
+      to: PlugCheckup,
+      init_opts:
+        PlugCheckup.Options.new(
+          json_encoder: Jason,
+          checks: ElixirBoilerplateHealth.checks(),
+          error_code: ElixirBoilerplateHealth.error_code(),
+          timeout: :timer.seconds(5),
+          pretty: false
+        )
+    )
+  end
+
+  plug(:match)
+  plug(:dispatch)
+
+  forward("/health", to: Health)
+
+  match(_, do: conn)
+end
